+++ mrmustard/lab_dev/transformations/__init__.py
@@ -16,5 +16,12 @@
 The classes representing transformations in quantum circuits.
 """
 
+from .amplifier import *
+from .attenuator import *
 from .base import *
-from .transformations import *
+from .bsgate import *
+from .dgate import *
+from .identity import *
+from .rgate import *
+from .s2gate import *
+from .sgate import *
