+++ mrmustard/lab_dev/transformations/dgate.py
@@ -1,0 +1,97 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+The class representing a displacement gate.
+"""
+
+from __future__ import annotations
+
+from typing import Optional, Sequence, Tuple, Union
+
+from .base import Unitary
+from ...physics.representations import Bargmann
+from ...physics import triples
+from ..utils import make_parameter, reshape_params
+
+__all__ = ["Dgate"]
+
+
+class Dgate(Unitary):
+    r"""
+    The displacement gate.
+
+    If ``x`` and/or ``y`` are iterables, their length must be equal to `1` or `N`. If their length is equal to `1`,
+    all the modes share the same parameters.
+
+    .. code-block ::
+
+        >>> import numpy as np
+        >>> from mrmustard.lab_dev import Dgate
+
+        >>> unitary = Dgate(modes=[1, 2], x=0.1, y=[0.2, 0.3])
+        >>> assert unitary.modes == [1, 2]
+        >>> assert np.allclose(unitary.x.value, [0.1, 0.1])
+        >>> assert np.allclose(unitary.y.value, [0.2, 0.3])
+
+    Args:
+        modes: The modes this gate is applied to.
+        x: The displacements along the `x` axis, which represents position axis in phase space.
+        y: The displacements along the `y` axis.
+        x_bounds: The bounds for the displacement along the `x` axis.
+        y_bounds: The bounds for the displacement along the `y` axis, which represents momentum axis in phase space.
+        x_trainable: Whether `x` is a trainable variable.
+        y_trainable: Whether `y` is a trainable variable.
+
+    .. details::
+
+        For any :math:`\bar{\alpha} = \bar{x} + i\bar{y}` of length :math:`N`, the :math:`N`-mode
+        displacement gate is defined by
+
+        .. math::
+            S = I_N \text{ and } r = \sqrt{2\hbar}\big[\text{Re}(\bar{\alpha}), \text{Im}(\bar{\alpha})\big].
+
+        Its ``(A,b,c)`` triple is given by
+
+        .. math::
+            A &= \begin{bmatrix}
+                    O_N & I_N\\
+                    I_N & O_N
+                \end{bmatrix} \\ \\
+            b &= \begin{bmatrix}
+                    \bar{\alpha} & -\bar{\alpha}^*
+                \end{bmatrix} \\ \\
+            c &= \text{exp}\big(-|\bar{\alpha}^2|/2\big).
+    """
+
+    short_name = "D"
+
+    def __init__(
+        self,
+        modes: Sequence[int] = None,
+        x: Union[float, Sequence[float]] = 0.0,
+        y: Union[float, Sequence[float]] = 0.0,
+        x_trainable: bool = False,
+        y_trainable: bool = False,
+        x_bounds: Tuple[Optional[float], Optional[float]] = (None, None),
+        y_bounds: Tuple[Optional[float], Optional[float]] = (None, None),
+    ) -> None:
+        super().__init__(modes_out=modes, modes_in=modes, name="Dgate")
+        xs, ys = list(reshape_params(len(modes), x=x, y=y))
+        self._add_parameter(make_parameter(x_trainable, xs, "x", x_bounds))
+        self._add_parameter(make_parameter(y_trainable, ys, "y", y_bounds))
+
+        self._representation = Bargmann.from_function(
+            fn=triples.displacement_gate_Abc, x=self.x, y=self.y
+        )
